+++ src/clientsocket.h
@@ -8,6 +8,7 @@
 #ifndef CLIENTSOCKET_H
 #define CLIENTSOCKET_H
 
+#include <QtGlobal>
 #include <QTcpSocket>
 #include <QObject>
 #include <QDebug>
@@ -40,13 +41,13 @@ public:
     // Return -1 if an error occured
     //         0 if no more data is available for reading
     //         otherwise the size of byte read by buffer
-    ssize_t read(char* buffer, size_t size);// read maximum size bytes from socket buffer
-    ssize_t read_line(char* buffer, size_t size); // read one line(end with \n) from socket buffer
-    ssize_t read_all(char* buffer);  // read all content of socket buffer
+    qint64 read(char* buffer, size_t size);// read maximum size bytes from socket buffer
+    qint64 read_line(char* buffer, size_t size); // read one line(end with \n) from socket buffer
+    qint64 read_all(char* buffer);  // read all content of socket buffer
 
     // Return -1 if an error occured
     //        otherwise the size of byte write to buffer
-    ssize_t write(char* const & content, size_t size); // write the content to the socket buffer
+    qint64 write(char* const & content, size_t size); // write the content to the socket buffer
 };
 
 #endif // CLIENTSOCKET_H
